@@ -39,22 +39,26 @@
     "not op_mini all"
   ],
   "dependencies": {
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mui/material": "^5.11.9",
+    "axios": "^1.3.3",
     "bootstrap": "4.3.1",
+    "boring-avatars": "^1.7.0",
     "classnames": "2.3.1",
     "headroom.js": "0.12.0",
     "moment": "2.29.4",
     "nouislider": "14.1.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-datetime": "3.1.1",
-    "react-dom": ">=16.0.0",
+    "react-dom": "^18.2.0",
+    "react-ion-icon": "^4.0.0",
     "react-ionicons": "^4.2.0",
-    "react-router-dom": "^5.3.4",
+    "react-router-dom": "^6.8.0",
     "react-scripts": "5.0.1",
     "reactstrap": "8.4.1",
-    "sass": "1.53.0"
-  },
-  "peerDependencies": {
-    "react-dom": ">=16.0.0"
+    "sass": "1.53.0",
+    "spinners-react": "^1.0.7"
   },
   "devDependencies": {
     "@types/googlemaps": "3.39.2",
